Conditionalize test

+++ tests/testthat/test-sgrender.R
@@ -74,15 +74,18 @@ test_that("sgrender2: Render output types works.", {
   expect_equal(file.exists(res$path), TRUE)
 
   # SVG
-  fp <- file.path(base_path, "output/test.svg")
+  if (capabilities("cairo")) {
+    fp <- file.path(base_path, "output/test.svg")
 
-  if (file.exists(fp)) {
-    file.remove(fp)
+    if (file.exists(fp)) {
+      file.remove(fp)
+    }
+
+    res <- sgrender(p, output_type = "svg", file_path = fp)
+
+    expect_equal(file.exists(res$path), TRUE)
+
   }
-
-  res <- sgrender(p, output_type = "svg", file_path = fp)
-
-  expect_equal(file.exists(res$path), TRUE)
 
   # TIF
   fp <- file.path(base_path, "output/test.tif")
